refactor: refactor wt to preserve workers results sequence'
