+++ src/wt/main.js
@@ -1,61 +1,43 @@
-import { dirname, join, resolve } from 'path';
+import { dirname, join } from 'path';
 import { fileURLToPath } from 'url';
 import { cpus } from 'os';
 import { Worker } from 'worker_threads';
 
-const statuses = {
-  RESOLVED: 'resolved',
-  ERROR: 'error',
-  PENDING: 'pending',
-  FULFILLED: 'fulfilled',
-  REJECTED: 'rejected',
-};
+import modifyWorkerOutput from '../utils/modifyWorkerOutput.mjs';
 
-const START_NUM = 10;
 const FILE_NAME = 'worker.js';
 const ERROR_MESSAGE = 'Worker error';
+let START_NUM = 10;
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 const performCalculations = async () => {
-  const coresNum = cpus().length;
-  const results = [];
+  const cores = cpus();
 
   const workerPath = join(__dirname, FILE_NAME);
 
-  for (let i = 0; i < coresNum; i++) {
-    const workerRes = await new Promise((resolve, reject) => {
-      const worker = new Worker(workerPath, {
-        workerData: START_NUM + i,
+  const workerMessages = await Promise.allSettled(
+    cores.map(() => {
+      return new Promise((resolve, reject) => {
+        const worker = new Worker(workerPath, {
+          workerData: START_NUM++,
+        });
+
+        worker.on('message', (message) => {
+          resolve(message);
+        });
+
+        worker.on('error', (error) => {
+          console.error(`${ERROR_MESSAGE}: ${error}`);
+          reject(error);
+        });
       });
-
-      worker.on('message', (message) => {
-        resolve({ status: statuses.RESOLVED, message });
-      });
-
-      worker.on('error', (error) => {
-        console.error(`${ERROR_MESSAGE}: ${error}`);
-        reject({ status: statuses.ERROR, data: null });
-      });
-    });
-
-    results.push(workerRes);
-  }
-
-  Promise.allSettled(results)
-    .then((results) => {
-      const formattedResults = results.map((result) =>
-        result.status === statuses.FULFILLED
-          ? { status: statuses.RESOLVED, data: result.value.message }
-          : { status: statuses.ERROR, data: null }
-      );
-
-      console.log(formattedResults);
     })
-    .catch((err) => {
-      throw err;
-    });
+  );
+
+  const res = modifyWorkerOutput(workerMessages);
+  console.log(res);
 };
 
 await performCalculations();
